Refactor out the slides into their separate files, bringing in along the styles.
