+++ src/baseStyles.js
@@ -1,0 +1,4 @@
+export const imageStyles = {
+    width: "100%",
+    height: "100%"
+};
