+++ src/resume.json
@@ -2,7 +2,7 @@
   "name": "Sang Bui",
   "phone": "714-388-8314",
   "email": "sangbui51@yahoo.com",
-  "about": "This is my personal page!!",
+  "about": "This is my personal resume page!!",
   "summary":
       "In 2011, I didn't know OOP, Java, C++, or game development tools. In 2013, I didn't know about JavaScript, Node or its ecosystem; web, iOS, or Android development. In 2015, I didn't know about React/Redux or functional programming. In 2017, I didn't know about Neural Nets or Reinforcement Learning. My biggest asset is not the languages or skills I had the pleasure of learning, but what I could learn.  Learning new programming paradigm, tools, and languages is my best skill."
     ,
@@ -18,20 +18,14 @@
       "degree": "Master of Science in Computer Science"
     }
   ],
-  "experience": [
+  "experiences": [
     {
       "company": "Northrop Grumman",
       "from": "June 2015",
       "to": "Present",
       "position": "Software Engineer",
       "description": [
-        "Code review for other developers.",
-        "Defined product requirements and creates high-level architectural specifications, ensuring feasibility and functionality for a full web application.",
-        "Architect a web application with React/Redux, python backend, Kerbereros Single Sign On Authentication system.",
-        "Created Base of Estimate on how many hours/cost future project will take.",
-        "Developed software and follows the company SDLC (Software Development Lifecycle), which includes Agile/Scrum sprint tasking and management using Jira/Gitlab",
-        "Worked on a Situational Awareness Web Application",
-        "Help created a Selenium testing suite."
+        "Code review for other developers. Defined product requirements and creates high-level architectural specifications, ensuring feasibility and functionality for a full web application. Architect a web application with React/Redux, python backend, Kerbereros Single Sign On Authentication system. Created Base of Estimate on how many hours/cost future project will take. Developed software and follows the company SDLC (Software Development Lifecycle), which includes Agile/Scrum sprint tasking and management using Jira/Gitlab. Worked on a Situational Awareness Web Application.  Help created a Selenium testing suite."
       ]
     }
